Create README.md
Car Sales system is an application that computerizes the conventional car sale procedure which we are aware of. This is a Java/HTML based application. This helps in managing data related to buyers and sellers of the cars. Business reports can also be generated and viewed.

This program handles  new and used vehicles inventory with minimum 10 cars. Every detail of a vehicle like manufacturer name, date of purchase, Kilo meters ran( in case of used vehicles), price, make/model, car description picture/s of the car etc.

If a user click to view details of a car the application displays the car details :eg
name, model, year build, price, description, picture, etc...

If a car is sitting in inventory for more than 120 days, there is an option to place a bid on the car.
Bidding cars can be sold at a discounted price up to 10% if and only if the dealership's purchase date is over 120 days.

The inventory  has minimum 3 cars with idle days of 120+.
The inventory has an option to replenish the inventory by adding pictures and details of a car.

The application to have a reporting section which displays information about all transactions made, including the person who purchase the car. 
This reporting page displays the order transaction by date starting with the most recent purchase.

A user is also  able to search for a type (model) car from the inventory.

The inventory has cars available to be sold. If a car has been sold the inventory reflects this behavior. 

Note:
This system has local storage of all sale transactions (save record to a  text file).
This text file is loaded on every application load. This means the application keeps a record of the transaction history even after it has been powered down.